Move configs from package and refine
